config: runtime: base: lava: add ability to append an upload template
Some jobs might need to push artifacts to an online storage, so they can
be retrieved and further processed as a follow-up task. This is
implemented in the main LAVA template by include an additional "test"
template, providing all the logic for packing and uploading the needed
artifacts.

As this likely requires some form of authentication, this change also
adds a `secrets` dictionary, providing an easy way to setup a token as
env var; please note, however, this relies on "personal tokens" setup in
LAVA, like is already done for authenticating notify callbacks. Failure
to do so will result in the token being leaked in the job definition!